Fix parsing of code blocks and spans
Parsing the contents of code blocks and spans is now handled at the
lexer level, instead of the parser doing most of the work. This makes it
much easier to handle formatting inside code spans/blocks, and fixes
various bugs that would result in input being swallowed in certain
cases.

Changelog: fixed